requirements.txt
asgiref==3.8.1
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
MarkupSafe==2.1.5
passlib==1.7.4
psycopg2-binary==2.9.9
PyJWT==1.7.1
python-dotenv==1.0.1
sqlparse==0.5.1
Werkzeug==3.0.3
